Partition number uint16_t -> uint32_t (rf_kernels)

+++ plugins/jml/randomforest_kernels.cc
@@ -1262,7 +1262,7 @@ DEFINE_STRUCTURE_DESCRIPTION_INLINE(PartitionSplit)
 }
 
 void updateBuckets(const std::vector<Feature> & features,
-                   std::vector<uint16_t> & partitions,
+                   std::vector<uint32_t> & partitions,
                    std::vector<std::vector<W> > & buckets,
                    std::vector<W> & wAll,
                    const std::vector<uint32_t> & bucketOffsets,
@@ -1435,7 +1435,7 @@ std::pair<std::vector<PartitionEntry>,
 splitPartitions(const std::vector<Feature> features,
                 const std::vector<int> & activeFeatures,
                 const std::vector<float> & decodedRows,
-                const std::vector<uint16_t> & partitions,
+                const std::vector<uint32_t> & partitions,
                 const std::vector<W> & w,
                 MappedSerializer & serializer)
 {
@@ -1722,7 +1722,7 @@ getPartitionSplits(const std::vector<std::vector<W> > & buckets,
 
 // Check that the partition counts match the W counts.
 void
-verifyPartitionBuckets(const std::vector<uint16_t> & partitions,
+verifyPartitionBuckets(const std::vector<uint32_t> & partitions,
                        const std::vector<W> & wAll)
 {
     using namespace std;
@@ -1757,7 +1757,7 @@ splitAndRecursePartitioned(int depth, int maxDepth,
                            const std::vector<Feature> & features,
                            const std::vector<int> & activeFeatures,
                            const std::vector<float> & decodedRows,
-                           const std::vector<uint16_t> & partitions,
+                           const std::vector<uint32_t> & partitions,
                            const std::vector<W> & wAll,
                            const DatasetFeatureSpace & fs,
                            FrozenMemoryRegionT<uint32_t> bucketMemory)
@@ -1835,7 +1835,9 @@ trainPartitionedRecursiveCpu(int depth, int maxDepth,
     // is in partition zero, but as we start to split, we end up
     // splitting them amongst many partitions.  Each partition has
     // its own set of buckets that we maintain.
-    std::vector<uint16_t> partitions(rowCount, 0);
+    // Note that these are indexes into the partition list, to
+    // enable sparseness in the tree.
+    std::vector<uint32_t> partitions(rowCount, 0);
 
     // What is our weight total for each of our partitions?
     std::vector<W> wAll = { wAllInput };
@@ -2578,7 +2580,7 @@ trainPartitionedEndToEndOpenCL(int depth, int maxDepth,
     // splitting them amongst many partitions.  Each partition has
     // its own set of buckets that we maintain.
 
-    size_t partitionMemoryUsage = sizeof(uint16_t) * rowCount;
+    size_t partitionMemoryUsage = sizeof(uint32_t) * rowCount;
     
     OpenCLMemObject clPartitions
         = context.createBuffer(CL_MEM_READ_WRITE, partitionMemoryUsage);
@@ -2587,7 +2589,7 @@ trainPartitionedEndToEndOpenCL(int depth, int maxDepth,
 
     // Before we use this, it needs to be zero-filled
     OpenCLEvent fillPartitions
-        = queue.enqueueFillBuffer(clPartitions, (uint16_t)0 /* pattern */,
+        = queue.enqueueFillBuffer(clPartitions, (uint32_t)0 /* pattern */,
                                   0 /* offset */, partitionMemoryUsage);
 
     allEvents.emplace_back("fillPartitions", fillPartitions);
@@ -2807,7 +2809,7 @@ trainPartitionedEndToEndOpenCL(int depth, int maxDepth,
         std::vector<PartitionSplit> debugPartitionSplitsCpu;
         std::vector<std::vector<W> > debugBucketsCpu;
         std::vector<W> debugWAllCpu;
-        std::vector<uint16_t> debugPartitionsCpu;
+        std::vector<uint32_t> debugPartitionsCpu;
 
         if (depth == maxDepth - 1 && false) {
             OpenCLEvent mapPartitionSplits;
@@ -2894,8 +2896,8 @@ trainPartitionedEndToEndOpenCL(int depth, int maxDepth,
         
             mapPartitions.waitUntilFinished();
 
-            const uint16_t * partitionsGpu
-                = reinterpret_cast<const uint16_t *>
+            const uint32_t * partitionsGpu
+                = reinterpret_cast<const uint32_t *>
                     (mappedPartitions.get());
 
             debugPartitionsCpu = { partitionsGpu, partitionsGpu + rowCount };
@@ -3229,8 +3231,8 @@ trainPartitionedEndToEndOpenCL(int depth, int maxDepth,
         
             mapPartitions.waitUntilFinished();
 
-            const uint16_t * partitionsGpu
-                = reinterpret_cast<const uint16_t *>
+            const uint32_t * partitionsGpu
+                = reinterpret_cast<const uint32_t *>
                     (mappedPartitions.get());
 
             for (size_t i = 0;  i < rowCount;  ++i) {
